+++ apps/front/src/App.tsx
@@ -28,10 +28,10 @@ import StreetViewWindow from "~/features/street-view/StreetViewWindow";
 import TabsControl from "./TabsControl";
 import { RFrameRateControl } from "pentatrion-geo";
 import { debug } from "~/config/constants";
-import { selectGeolocationEnabled } from "~/features/geolocation/geolocationSlice";
 import GeolocationMap from "~/features/geolocation/GeolocationMap";
 import IsochroneControl from "./features/isochrone/IsochroneControl";
 import { selectIsochroneReferenceFeature } from "./features/isochrone/isochroneSlice";
+import { selectGeolocationStatus } from "./features/geolocation/geolocationSlice";
 
 function handleAfterMapInstanciation(map: Map) {
   map.loadImage("/assets/graphics/icons/arrow.png").then((img) => {
@@ -64,7 +64,7 @@ function App() {
   const terrain = useAppSelector(selectTerrain);
   const streetView = useAppSelector(selectStreetView);
   const tab = useAppSelector(selectTab);
-  const geolocationEnabled = useAppSelector(selectGeolocationEnabled);
+  const geolocationEnabled = useAppSelector(selectGeolocationStatus) === "on";
   const isochroneReferenceFeature = useAppSelector(selectIsochroneReferenceFeature);
 
   const [uncontrolledStyle, setUncontrolledStyle] = useState<StyleSpecification | string>({