+++ apps/front/src/features/geolocation/geolocationSlice.ts
@@ -2,12 +2,12 @@ import { PayloadAction, createSlice } from "@reduxjs/toolkit";
 import { AppDispatch, RootState } from "~/store";
 
 type ErrorCode = number | 1 | 2 | 3 | null;
+type Status = "off" | "waiting" | "on" | "error";
 
-type GeolocationState = {
-  enabled: boolean;
+export type GeolocationState = {
+  status: Status;
   coords: [number, number] | null;
   accuracy: number | null;
-  status: "off" | "waiting" | "on" | "error";
 
   // PERMISSION_DENIED: 1;
   // POSITION_UNAVAILABLE: 2;
@@ -19,13 +19,13 @@ type GeolocationState = {
 };
 
 const initialState: GeolocationState = {
-  enabled: false,
+  status: "off",
   coords: null,
   accuracy: null,
-  status: "off",
   errorCode: null,
-  showAccuracyCircle: true,
+
   lockCamera: true,
+  showAccuracyCircle: true,
 };
 
 let watchId: number | null = null;
@@ -40,34 +40,33 @@ const geolocationSlice = createSlice({
     lockCameraChanged(state, action: PayloadAction<boolean>) {
       state.lockCamera = action.payload;
     },
-    enabledChanged(state, action: PayloadAction<boolean>) {
-      const enabled = action.payload;
-      if (state.enabled === enabled) {
+    statusChanged(state, action: PayloadAction<Status>) {
+      const status = action.payload;
+      if (state.status === status) {
         return;
       }
 
-      if (!state.enabled) {
-        state.status = "waiting";
-      } else {
-        state.status = "off";
-      }
+      state.status = status;
+      state.accuracy = null;
       state.coords = null;
-      state.enabled = action.payload;
       state.errorCode = null;
     },
     positionUpdated(state, action: PayloadAction<{ coords: [number, number]; accuracy: number }>) {
       const { coords, accuracy } = action.payload;
       console.log("positionUpdated", coords, accuracy);
+
+      state.status = "on";
       state.coords = coords;
       state.accuracy = accuracy;
-      state.status = "on";
       state.errorCode = null;
     },
     positionErrored(state, action: PayloadAction<ErrorCode>) {
+      console.log("positionErrored", action.payload);
+
       state.status = "error";
-      state.errorCode = action.payload;
       state.coords = null;
-      console.log("positionErrored", state.errorCode);
+      state.accuracy = null;
+      state.errorCode = action.payload;
     },
   },
 });
@@ -77,17 +76,18 @@ export default geolocationSlice.reducer;
 export const {
   showAccuracyCircleChanged,
   lockCameraChanged,
-  enabledChanged,
+  statusChanged,
   positionUpdated,
   positionErrored,
 } = geolocationSlice.actions;
 
 export const selectGeolocationCoords = (state: RootState) => state.geolocation.coords;
-export const selectGeolocationEnabled = (state: RootState) => state.geolocation.enabled;
+export const selectGeolocationStatus = (state: RootState) => state.geolocation.status;
 export const selectGeolocation = (state: RootState) => state.geolocation;
 
 export const activationChanged = (enabled: boolean) => (dispatch: AppDispatch) => {
-  dispatch(enabledChanged(enabled));
+  dispatch(statusChanged(enabled ? "waiting" : "off"));
+
   if (enabled) {
     if (watchId === null) {
       watchId = navigator.geolocation.watchPosition(
@@ -102,7 +102,6 @@ export const activationChanged = (enabled: boolean) => (dispatch: AppDispatch) =
         ({ code }) => {
           dispatch(positionErrored(code));
           if (code === GeolocationPositionError.PERMISSION_DENIED) {
-            dispatch(enabledChanged(false));
             if (watchId) {
               navigator.geolocation.clearWatch(watchId);
               watchId = null;