+++ apps/front/src/features/search/FeatureInfos.tsx
@@ -6,6 +6,7 @@ import { coordsUnitChanged, selectCoordsUnit, tabChanged } from "~/store/mapSlic
 import { useNotification } from "pentatrion-design/redux";
 import { useT } from "talkr";
 import { useCallback, useState } from "react";
+
 import { directionWayPointsAddedFromSearch } from "~/features/direction/directionSlice";
 import { referenceFeatureChanged } from "../isochrone/isochroneSlice";